Retrieve Committee Data from the Austrian Parliament API

Get data on the committeses of the Austrian Parliament. Data includes session dates, agendas, meeting overviews, and member lists. The function mirrors the search functionality of the Austrian Parliament's website for committees here.

Usage

get_committees(
  search_string = NULL,
  institution = NULL,
  legis_period = NULL,
  permanent = NULL,
  include_subcommittees = NULL,
  echo = NULL
)

Arguments

search_string

A character string for free text search within committee records. Optional.

institution

A character string specifying the institution ("Bundesrat" or "Nationalrat"). This parameter is mandatory and is internally encoded to "BR" or "NR" respectively.

legis_period

A character vector representing one or more legislative period codes. Each element is validated using an internal helper function.

permanent

A logical flag indicating whether only permanent committees should be queried. When set to TRUE, the corresponding API parameter is adjusted accordingly. Optional.

include_subcommittees

A logical flag to indicate whether subcommittees should be included in the search results. Note that if permanent is TRUE, subcommittee search is disallowed.

echo

Logical. If TRUE, the function prints the used search parametes and the url to the pertaining search results on website of the Austrian Parlament.

Value

A data frame containing the committee data if successful. If no results match the criteria, a message is printed and NULL is returned.

Details

The function constructs a JSON request body by assembling the provided parameters (after validation and necessary encoding) and sends a POST request to the API endpoint. The API response is parsed, and the data is converted into a data frame with cleaned column names.

Examples

if (FALSE) { # \dontrun{
  # Retrieve data for the Nationalrat with a specified legislative period and permanent flag.
  df <- get_committees(
    search_string = "Korruption",
    institution = "Nationalrat",
    legis_period = 27
  )
  if (!is.null(df)) {
    print(df)
  }
} # }