Search for Parliamentary Items

get_items searches for items ('Verhandlungsgegenstände') that are or were subject to negotiations in the Austrian National Council ('Nationalrat') or Federal Council ('Bundesrat'). This function mirrors the search functionality offered on the Austrian Parliament's website. See here.

Usage

get_items(
  search_string = NULL,
  topic = NULL,
  institution = "Nationalrat",
  legis_period = NULL,
  date_start = NULL,
  date_end = NULL,
  item = NULL,
  doc_type = NULL,
  person = NULL,
  number = NULL,
  keyword = NULL,
  eurovoc = NULL,
  parl_group = NULL,
  parl_group_names_standard = FALSE,
  echo = TRUE
)

Arguments

topic

Character vector or NULL. Specifies the topic(s) to search for. See 'Details' for possible values. Default is NULL.

institution

Character string. Either "Nationalrat" (National Council) or "Bundesrat" (Federal Council). Default is "Nationalrat".

legis_period

Character vector or NULL. Specifies the legislative period(s) to search in. See 'Details' for possible values. Default is NULL.

date_start

Character string. Start date for the search period in format "dd-mm-yyyy". Default is NULL.

date_end

Character string. End date for the search period in format "dd-mm-yyyy". Default is NULL.

item

Character vector or NULL. Specifies the type(s) of parliamentary item(s) to search for. See 'Details' for possible values. Default is NULL.

doc_type

Character vector or NULL. Specifies the type of parliamentary item(s) to search for. See 'Details' for possible values. Default is NULL.

person

Character string or NULL. Name of a person to search for (family name, optionally followed by first name). Default is NULL.

number

Character string or NULL. Specific item number to search for. Default is NULL.

keyword

Character vector or NULL. Keyword(s) to search for. Default is NULL.

eurovoc

Character vector or NULL. EuroVoc term(s) to search for. Default is NULL.

parl_group

Character vector or NULL. Parliamentary group(s) to search for. Default is NULL. Combine multiple groups in a vector, i.e. c("SPÖ", "ÖVP"). See 'Details.'

parl_group_names_standard

Logical. If TRUE, the function expands and standardizes parliamentary group names. Default is FALSE. See 'Details.'

echo

Logical. If TRUE, the function prints the used search parametes and the url to the pertaining search results on website of the Austrian Parlament.

Value

A data frame containing the search results. If no results are found, the function returns NULL and displays a message.

Details

Topic ('Thema')

Possible values for topic include:

"Arbeit" (work)

"Außenpolitik" (foreign policy)

"Bildung" (education)

"Budget und Finanzen" (budget and finance)

"Europäische Union" (European Union)

"Familie und Generationen" (family and generations)

"Frauen und Gleichbehandlung" (women and equality)

"Gesundheit und Ernährung" (health and nutrition)

"Information und Medien" (information and media)

"Inneres und Recht" (interior and law)

"Innovation, Technologie und Forschung" (innovation, technology and research)

"Klima, Umwelt und Energie" (climate, environment and energy)

"Kultur" (culture)

"Land- und Forstwirtschaft" (agriculture and forestry)

"Landesverteidigung" (national defense)

"Parlament und Demokratie" (parliament and democracy)

"Soziales" (social affairs)

"Sport" (sports)

"Verkehr und Infrastruktur" (transport and infrastructure)

"Wirtschaft" (economy)

legis_period (Gesetzgebungsperiode)

legis_period specifies the legislative period(s). Can be one or more of the following value(s):

numbers indicating the relevant period(s),

"Provisorische Nationalversammlung" (Provisional National Assembly, 1918-1919),

and/or "Konstituierende Nationalversammlung" (Constituent National Assembly, 1919-1920).

item (Gegenstand)

Possible values for item include:

"ASEU" (Aktuelle Europastunden, Current European Hours)

"AS" (Aktuelle Stunden, Current Hours)

"J_JPR_M" (Anfragen, Inquiries)

"ANTR" (Anträge, Motions)

"US" (Anträge/Verlangen auf Untersuchungsausschuss, Motions/Requests for Investigative Committee)

"AUB" (Ausschussberichte, Committee Reports)

"AB_ABPR_ABM" (Beantwortungen, Answers)

"III" (Berichte an den Nationalrat, Reports to the National Council)

"BNR" (Beschlüsse, Resolutions)

"BI" (Bürgerinitiativen, Citizen Initiatives)

"E" (Einsprüche des Bundesrates, Objections of the Federal Council)

"EBR" (Entschließungen, Resolutions)

"EU" (EU betreffende Vorlagen und Beschlüsse, EU-related Proposals and Resolutions)

"FS" (Fragestunden, Question Time)

"GO" (Geschäftsbehandlung, Rules of Procedure)

"GABR" (Gesetzesanträge des Bundesrates, Legislative Proposals of the Federal Council)

"GABR13" (Gesetzesanträge von einem Drittel des BR, Legislative Proposals of One-Third of the Federal Council)

"IMM" (Immunitätsangelegenheiten, Immunity Matters)

"KOMM" (Kommuniqués, Communiqués)

"PET" (Petitionen, Petitions)

"RGER" (Regierungserklärungen, Government Statements)

"RV" (Regierungsvorlagen (Gesetze), Government Bills (Laws))

"RVS" (Staatsverträge, State Treaties)

"TRAU" (Trauerkundgebungen, Condolence Expressions)

"RVS15" (Vereinbarungen gemäß Art. 15a B-VG, Agreements Pursuant to Article 15a of the Federal Constitutional Law)

"VOLKBG" (Volksbegehren, Popular Initiatives)

"W" (Wahlen, Elections)

doc_type (Art des Antrages)

If item is "ANTR", the permissible values for doc_type depend on the institution of interest. Possible values for doc_type if institution=="Nationalrat" ('National Council'):

A: Selbständiger Antrag

A(E): Selbständiger Entschließungsantrag

AA: Abänderungsantrag

AEA: Selbständiger Ausschuss-Entschließungsantrag

AMIN: Selbständiger Antrag - Ministeranklage

ARH2: Verlangen auf Gebarungsüberprüfung durch den Rechnungshof

AVB: Antrag auf Volksbefragung

BUA: Bericht und Antrag

UEA: Unselbständiger Entschließungsantrag #BUG

UEA: Misstrauensantrag

URH2: Verlangen auf Gebarungsüberprüfung durch den Ständigen UA des Rechnungshofausschusses

Possible values for doc_type if institution=="Bundesrat" ('Federal Council'):

AA-BR: Abänderungsanträge

A-BR: Selbständiger Antrag Bundesrat

A(E): Selbständiger Entschließungsantrag Bundesrat

AEA-BR: Selbständige Entschließungsanträge von Ausschüssen

UEA-BR: Unselbständige Anträge

Possible values for doc_type if item=="BNR":

BNR: Beschluss

BS: Sonstiger Beschluss

BSE: Beschluss-EU

BSESM: Beschluss-ESM

BS-BR: Sonstiger Beschluss (only if instituition=="Bundesrat")

EuroVoc

EuroVoc is an international thesaurus developed primarily for use within the EU. It enables searches using standardized keywords across Europe. The EuroVoc search is supported for all negotiation items from the 20th legislative period onwards.

Parlamentary Group (Klub/Fraktion)

parl_group specifies the parliamentary group(s) to search for. The API of the Austrian Parliament accepts only specific abbreviations for each group:

BZÖ Bündnis Zkunft Österreich

CSP Crhistlichsoziale Partei

DnP Deutsche Nationalpartei

F Freiheitliche Partei Österreichs

F-BZÖ Freiheitliche Partei Österreichs - Bündnis Zukunft Österreich

FPÖ Freiheitliche Partei Österreichs

GdP Großdeutsche Volkspartei

GRÜNE Die Grünen - Die Grüne Alternative

HB Heimatblock

JETZT Jetzt - Liste Pilz

Konvent

KPÖ Kommunistische Partei Österreichs

KuL

L

LB

LBd Landbund für Österreich

NEOS NEOS - Das Neue Österreich

NEOS-LIF NEOS - Liberales Forum

NSDAP Nationalsozialistische Deutsche Arbeiterpartei

NWB Nationaler Wirtschaftsblck und Landbund

OF

OK Ohne Klub

ÖVP Österreichische Volkspartei

PILZ Liste Pilz

SdP

SPÖ Sozialistische/Sozialdemokratische Partei Österreichs

STRONACH Team Stronach

VO Wahlgemeinschafft Österreichische Volksopposition

WdU Wahlpartei der Unabhängigen (VdU, Verband der Unabhängigen)

Note

The API currently has a limitation of returning a maximum of 100,000 rows.

Examples

if (FALSE) { # \dontrun{
get_items(topic="Europäische Union", legis_period=28)
} # }