Get Members of Parliament

Retrieves information about Members of Parliament from the Austrian Parliament database. This function mirrors the search functionality 'Parlamentarier:innen ab 1918' from the Austrian Parliament website.

Usage

get_mps(
  institution = c("all", "Bundesrat", "Nationalrat"),
  gender = c("all", "female", "male"),
  legis_period = "all",
  party = c("all", "BP", "BZÖ", "BAP", "CSP", "Grüne", "FPÖ", "GdP", "HB", "KuL",
    "KPÖ", "LBd", "L", "LB", "PILZ", "NWB", "NEOS", "ÖVP", "SdP", "SPÖ", "STRONACH",
    "VO", "WdU"),
  parl_group = c("all", "LBd", "CSP", "GRÜNE", "SPÖ", "F-BZÖ", "GdP", "F", "FPÖ",
    "KuL", "VO", "WdU", "LB", "NEOS-LIF", "PILZ", "NEOS", "OK", "HB", "KPÖ", "ÖVP",
    "BZÖ", "JETZT", "L", "STRONACH", "NWB", "SdP"),
  electoral_district = c("all", "Bundeswahlvorschlag", "Burgenland", "Kärnten",
    "Niederösterreich", "Oberösterreich", "Salzburg", "Steiermark", "Tirol",
    "Vorarlberg", "Wien", "Wien Nord"),
  state = NULL,
  presidents_only = FALSE,
  make_unique = FALSE
)

Arguments

institution

Chamber of Parliament. One of "all", "Bundesrat", or "Nationalrat"

gender

Gender filter. One of "all", "female", or "male"

legis_period

Legislative period. Can be "all", a numeric value, "Provisorische Nationalversammlung", or "Konstituierende Nationalversammlung"

party

Political party filter. See details for possible values.

parl_group

Parliamentary group filter

electoral_district

Electoral district filter

state

State filter

presidents_only

Logical. If TRUE, returns only presidents. Default is FALSE

make_unique

Logical. If TRUE, returns unique entries. Default is FALSE

Value

A dataframe containing information about Members of Parliament including:

name

Name of the MP

fraktion

Parliamentary group

bundesland

Federal state

gesetzgebungsperioden

Legislative periods

And additional columns depending on the query parameters

Details

Available party values include: "all", "BP", "BZÖ", "BAP", "CSP", "Grüne", "FPÖ", "GdP", "HB", "KuL", "KPÖ", "LBd", "L", "LB", "PILZ", "NWB", "NEOS", "ÖVP", "SdP", "SPÖ", "STRONACH", "VO", "WdU"

Examples

if (FALSE) { # \dontrun{
# Get all MPs from the current legislative period
mps <- get_mps(institution = "Nationalrat", legis_period = "27")

# Get female MPs from a specific party
female_mps <- get_mps(gender = "female", party = "SPÖ")
} # }