Get Event Data from Parlament API

This function retrieves event data based on search parameters from the Parlament API. It largely mirrors the search functionality on the Austrian Parliament website at this page.

Usage

get_events(
  search_string = NULL,
  institution = "Nationalrat",
  event_type = NULL,
  place = NULL,
  date_start = NULL,
  date_end = NULL,
  echo = TRUE
)

Arguments

search_string

Optional character string used to further filter the search results.

institution

Character specifying the institution to query. Must be one of "Bundesrat" or "Nationalrat".

event_type

Optional character string indicating the event type. See Details.

place

Optional character string to filter events by location. Must match one of the predefined choices.

date_start

Optional character string representing the start date in "dd-mm-yyyy" format.

date_end

Optional character string representing the end date in "dd-mm-yyyy" format.

echo

Logical indicating; prints used search parameters, number of hits, and link to results on website of parliament.

Value

A data frame containing event details if results are found. If no events match the criteria, the function prints a message and returns NULL.

Details

Event type

Allowed event types include:

"Plenarsitzung"

"Ausschusssitzung oder Ausschuss"

"Besuch einer Plenarsitzung"

"Demokratiebildung"

"Fest-/Gedenksitzung"

"Führung"

"Internationales"

"Klubveranstaltung"

"Konferenz"

"Parlamentarische Enquete"

"Pressekonferenz"

"Sitzung der Bundesversammlung"

"Sonstiger Termin"

"Veranstaltung".

Examples

if (FALSE) { # \dontrun{
  # Example: Retrieve events from the Nationalrat between "01-01-2022" and "31-01-2022"
  events <- get_events(
    search_string = "budget",
    institution = "Nationalrat",
    event_type = "Plenarsitzung",
    place = "Nationalratssaal",
    date_start = "01-01-2022",
    date_end = "31-12-2022",
    echo = TRUE
  )
} # }