Get Current Members of Parliament

Fetches current members of parliament based on provided search criteria. Mirrors the search functionality on the Austrian Parliament website at this page.

Usage

get_mps_current(
  institution = NULL,
  gender = "all",
  position = NULL,
  parl_group = NULL,
  party = NULL,
  state = NULL,
  electoral_district = NULL,
  postal_code = NULL,
  echo = TRUE
)

Arguments

institution

Character. The parliamentary institution, accepted values are "Nationalrat" or "Bundesrat".

gender

Character. Gender filter to apply; options are "all", "female", or "male". Default is "all".

position

Character. Position filter with acceptable values:

"ALLE": Alle (all members of parliament)

"1PNR": Präsident des Nationalrates (President of the National Council)

"2PNR": Zweiter Präsident des Nationalrates (Second President of the National Council)

"3PNR": Dritter Präsident des Nationalrates (Third President of the National Council)

"PRAES": Präsidialkonferenz (Presidential Conference)

"ZON": Ordner des Nationalrates (Regulators of the National Council)

"ZSN": Schriftführer des Nationalrates (Secretary of the National Council)

parl_group

Character. Parliamentary group filter. Acceptable values include "all", "LBd", "CSP", "GRÜNE", "SPÖ", "F-BZÖ", "GdP", "F", "FPÖ", "KuL", "VO", "WdU", "LB", "NEOS-LIF", "PILZ", "NEOS", "OK", "HB", "KPÖ", "ÖVP", "BZÖ", "JETZT", "L", "STRONACH", "NWB", "SdP".

party

Character. Political party filter with acceptable values such as "all", "BP", "BZÖ", "BAP", "CSP", "Grüne", "FPÖ", "GdP", "HB", "KuL", "KPÖ", "LBd", "L", "LB", "PILZ", "NWB", "NEOS", "ÖVP", "SdP", "SPÖ", "STRONACH", "VO", "WdU".

state

Character. (Reserved) Currently not utilized.

electoral_district

Character. Electoral district filter. Accepted values include region names such as "all", "Bundeswahlvorschlag", "Burgenland", "Kärnten", "Niederösterreich", "Oberösterreich", "Salzburg", "Steiermark", "Tirol", "Vorarlberg", "Wien", and various sub-regions (e.g., "Burgenland Nord", "Burgenland Süd", etc.).

Value

A data frame containing the list of current members of parliament that match the search criteria.

Examples

if (FALSE) { # \dontrun{
  df_members <- get_mps_current(
    institution = "Nationalrat",
    gender = "female",
    party = "SPÖ",
    electoral_district = "Wien"
  )

  if (!is.null(df_members)) {
    print(df_members)
  }
} # }