Get Participation Data from Austrian Parliament

This function retrieves participation data from the Austrian Parliament's API based on various filter criteria.

Usage

get_participation(
  topic = NULL,
  legis_period = NULL,
  active = NULL,
  review_type = NULL,
  initiative_type = NULL
)

Arguments

topic

(Themen)Character vector. Optional. Specifies the topic(s) of interest. Must be one or more of: "Arbeit", "Außenpolitik", "Bildung", "Budget und Finanzen", "Europäische Union", "Familie und Generationen", "Frauen und Gleichbehandlung", "Gesundheit und Ernährung", "Information und Medien", "Inneres und Recht", "Innovation, Technologie und Forschung", "Klima, Umwelt und Energie", "Kultur", "Land- und Forstwirtschaft", "Landesverteidigung", "Parlament und Demokratie", "Soziales", "Sport", "Verkehr und Infrastruktur", "Wirtschaft".

legis_period

(Gesetzgebungsperiode)Character vector. Optional. Specifies the legislative period(s).

active

(Aktuelle Beteiligung) Character. Optional. If "J", only includes current participations.

review_type

(Gegenstand) Character vector. Optional. Specifies the type of review. Must be one or more of: "RGES" (Gesetzesinitiativen), "ME" (Ministerialentwürfe), "BI" (Bürgerinitiativen), "PET" (Petitionen), "SN" (Stellungnahmen).

initiative_type

(Art der Gesetzesinitiative) Character vector. Optional. Specifies the type of legislative initiative. Must be one or more of: "EBR" (Einsprüche des Bundesrats), "GABR" (Gesetzesanträge von Abgeordneten), "A" (Gesetzesanträge von Abgeordneten), "GABR13" (Gesetzesanträge von 1/3 der Mitglieder des Bundesrats), "RV" (Regierungsvorlagen), "VOLKBG" (Volksbegehren).

department

(Ressort) Character vector. Optional. Only available if `review_type=='ME' (Ministerialentwürfe). Specifies the administrative unit which initiated the draft law. Must be one or more of: "BMJ (BM f. Justiz)", "BMK (BM f. Klimaschutz, Umwelt, Energie, Mobilität, Innovation u. Technologie)", "BMNT (BM f. Nachhaltigkeit u. Tourismus)", "BMLV (BM f. Landesverteidigung)", "BMVRDJ (BM f. Verfassung, Reformen, Deregulierung u. Justiz)", "BML (BM f. Land- u. Forstwirtschaft, Regionen u. Wasserwirtschaft)", "BMLRT (BM f. Landwirtschaft, Regionen u. Tourismus)", "BMSGPK (BM f. Soziales, Gesundheit, Pflege u. Konsumentenschutz)", "BMI (BM f. Inneres)", "BMF (BM f. Finanzen)", "BMDW (BM f. Digitalisierung u. Wirtschaftsstandort)", "BMBWF (BM f. Bildung, Wissenschaft u. Forschung)", "BMAW (BM f. Arbeit u. Wirtschaft)", "BMASGK (BM f. Arbeit, Soziales, Gesundheit u. Konsumentenschutz)", "BMAFJ (BM f. Arbeit, Familie u. Jugend)", "BMA (BM f. Arbeit)", "BMEIA (BM f. europäische u. internationale Angelegenheiten)", "BMKÖS (BM f. Kunst, Kultur, öffentlichen Dienst u. Sport)", "BMFFIM (Büro der Bundesministerin f. Frauen, Familie, Integration u. Medien)", "BMEUV (

Value

A data frame containing the participation data, or NULL if no results are found.

Details

This function sends a request to the Austrian Parliament's API to retrieve participation data based on the provided filter criteria. It performs input validation for each parameter and constructs the API request accordingly.

Examples

# Get participation data for the topic "Bildung"
education_data <- get_participation(topic = "Bildung")
#> [1] 14594

# Get participation data for multiple topics and legislative periods
multi_data <- get_participation(
  topic = c("Arbeit", "Soziales"),
  legis_period = c("27", "26"),
  review_type = "RGES"
)
#> [1] 467